appname/requirements.txt
asgiref==3.2.7
Babel==2.8.0
bootstrap4==0.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-allauth==0.41.0
django-autoslug==1.9.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==12.0.3
django-braces==1.14.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-payline==0.11
django-phonenumber-field==4.0.0
django-verified-email-field==1.6.1
django-widget-tweaks==1.4.8
idna==2.9
misaka==2.1.1
oauthlib==3.1.0
phonenumbers==8.12.4
Pillow==7.1.1
pycparser==2.20
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
stripe==2.48.0
suds-py3==1.4.1.0
urllib3==1.25.9
